weather-extractor-rest/pyproject.toml

[tool.poetry]
name = "met_extractor"
version = "1.0.0"
description = "A reimplementation of the MET extractor using the Cognite REST Extension for extractor-utils"
authors = ["Mathias Lohne <mathias.lohne@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["cognite"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
cognite-sdk-core = "^2.39.1"
cognite-extractor-utils = "^2.0.0"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.7"
macholib = {version = "^1.14", platform = "darwin"}             # Used by pyinstaller pn Mac OS
pywin32-ctypes = {version = "^0.2.0", platform = "win32"}       # Used by pyinstaller on Windows
pefile = "^2019.4.18"                                           # Used by pyinstaller on Windows
mypy = "^0.931"
flake8 = "^4.0.1"
black = "^21.12b0"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[tool.poetry.scripts]
met_extractor = "met_extractor.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
